requirements.txt
streamlit 
pytesseract 
googletrans==4.0.0-rc1 
gtts 
pillow 
python-magic
SpeechRecognition